#77 Adjust selects and fields

+++ _source/LigaCancer/Models/FormModel/FamilyMemberFormModel.cs
@@ -25,6 +25,6 @@ namespace LigaCancer.Models.FormModel
         [Display(Name = "Renda mensal")]
         [Range(0, 1000000.00, ErrorMessage = "Insira um valor de 0.00 até 1,000,000")]
         [RegularExpression(@"[0-9]{0,5}\.?[0-9]{1,2}", ErrorMessage = "Insira um valor válido!")]
-        public decimal? MonthlyIncome { get; set; }
+        public double? MonthlyIncome { get; set; }
     }
 }
